+++ ewm-main/src/main/java/ru/practicum/explorewithme/controller/ewmadmin/AdminCompilationController.java
@@ -5,9 +5,9 @@ import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.explorewithme.dto.in.create.NewCompilationDto;
-import ru.practicum.explorewithme.dto.in.update.UpdateCompilationRequest;
-import ru.practicum.explorewithme.dto.out.CompilationDto;
+import ru.practicum.explorewithme.dto.input.create.NewCompilationDto;
+import ru.practicum.explorewithme.dto.input.update.UpdateCompilationRequest;
+import ru.practicum.explorewithme.dto.output.CompilationDto;
 import ru.practicum.explorewithme.service.compilation.AdminCompilationService;
 
 import javax.validation.Valid;
