+++ ewm-main/src/main/java/ru/practicum/explorewithme/controller/ewmprivate/UserEventController.java
@@ -5,13 +5,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.checkerframework.checker.index.qual.Positive;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.explorewithme.dto.in.create.NewEventDto;
-import ru.practicum.explorewithme.dto.in.update.EventRequestStatusUpdateRequest;
-import ru.practicum.explorewithme.dto.in.update.UpdateEventUserRequest;
-import ru.practicum.explorewithme.dto.out.EventFullDto;
-import ru.practicum.explorewithme.dto.out.EventRequestStatusUpdateResult;
-import ru.practicum.explorewithme.dto.out.ParticipationRequestDto;
-import ru.practicum.explorewithme.dto.out.outshort.EventShortDto;
+import ru.practicum.explorewithme.dto.input.create.NewEventDto;
+import ru.practicum.explorewithme.dto.input.update.EventRequestStatusUpdateRequest;
+import ru.practicum.explorewithme.dto.input.update.UpdateEventUserRequest;
+import ru.practicum.explorewithme.dto.output.EventFullDto;
+import ru.practicum.explorewithme.dto.output.EventRequestStatusUpdateResult;
+import ru.practicum.explorewithme.dto.output.ParticipationRequestDto;
+import ru.practicum.explorewithme.dto.output.outshort.EventShortDto;
 import ru.practicum.explorewithme.service.event.PrivateEventService;
 import ru.practicum.explorewithme.service.request.PrivateRequestService;
