+++ ewm-main/src/main/java/ru/practicum/explorewithme/controller/ewmpublic/CompilationController.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.checkerframework.checker.index.qual.Positive;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.explorewithme.dto.out.CompilationDto;
+import ru.practicum.explorewithme.dto.output.CompilationDto;
 import ru.practicum.explorewithme.service.compilation.PublicCompilationService;
 
 import java.util.List;
