+++ ewm-main/src/main/java/ru/practicum/explorewithme/service/category/CategoryServiceImpl.java
@@ -6,8 +6,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.explorewithme.dto.in.create.NewCategoryDto;
-import ru.practicum.explorewithme.dto.out.CategoryDto;
+import ru.practicum.explorewithme.dto.input.create.NewCategoryDto;
+import ru.practicum.explorewithme.dto.output.CategoryDto;
 import ru.practicum.explorewithme.exeption.NotFoundException;
 import ru.practicum.explorewithme.exeption.NotMeetRulesException;
 import ru.practicum.explorewithme.model.Category;
