+++ ewm-main/src/main/java/ru/practicum/explorewithme/service/compilation/AdminCompilationService.java
@@ -1,8 +1,8 @@
 package ru.practicum.explorewithme.service.compilation;
 
-import ru.practicum.explorewithme.dto.in.create.NewCompilationDto;
-import ru.practicum.explorewithme.dto.in.update.UpdateCompilationRequest;
-import ru.practicum.explorewithme.dto.out.CompilationDto;
+import ru.practicum.explorewithme.dto.input.create.NewCompilationDto;
+import ru.practicum.explorewithme.dto.input.update.UpdateCompilationRequest;
+import ru.practicum.explorewithme.dto.output.CompilationDto;
 
 public interface AdminCompilationService {
