+++ ewm-main/src/main/java/ru/practicum/explorewithme/service/compilation/CompilationServiceImpl.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import ru.practicum.explorewithme.dto.in.create.NewCompilationDto;
-import ru.practicum.explorewithme.dto.in.update.UpdateCompilationRequest;
-import ru.practicum.explorewithme.dto.out.CompilationDto;
+import ru.practicum.explorewithme.dto.input.create.NewCompilationDto;
+import ru.practicum.explorewithme.dto.input.update.UpdateCompilationRequest;
+import ru.practicum.explorewithme.dto.output.CompilationDto;
 import ru.practicum.explorewithme.exeption.NotFoundException;
 import ru.practicum.explorewithme.model.Compilation;
 import ru.practicum.explorewithme.model.Event;
