+++ ewm-main/src/main/java/ru/practicum/explorewithme/service/request/RequestServiceImpl.java
@@ -3,9 +3,9 @@ package ru.practicum.explorewithme.service.request;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.practicum.explorewithme.dto.in.update.EventRequestStatusUpdateRequest;
-import ru.practicum.explorewithme.dto.out.EventRequestStatusUpdateResult;
-import ru.practicum.explorewithme.dto.out.ParticipationRequestDto;
+import ru.practicum.explorewithme.dto.input.update.EventRequestStatusUpdateRequest;
+import ru.practicum.explorewithme.dto.output.EventRequestStatusUpdateResult;
+import ru.practicum.explorewithme.dto.output.ParticipationRequestDto;
 import ru.practicum.explorewithme.exeption.NotFoundException;
 import ru.practicum.explorewithme.exeption.NotFoundListException;
 import ru.practicum.explorewithme.exeption.NotMeetRulesException;
