+++ ewm-main/src/main/java/ru/practicum/explorewithme/service/user/AdminUserService.java
@@ -1,7 +1,7 @@
 package ru.practicum.explorewithme.service.user;
 
-import ru.practicum.explorewithme.dto.in.create.NewUserRequestDto;
-import ru.practicum.explorewithme.dto.out.UserDto;
+import ru.practicum.explorewithme.dto.input.create.NewUserRequestDto;
+import ru.practicum.explorewithme.dto.output.UserDto;
 
 import java.util.List;
