+++ ewm-main/src/main/java/ru/practicum/explorewithme/service/user/UserServiceImpl.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import ru.practicum.explorewithme.dto.in.create.NewUserRequestDto;
-import ru.practicum.explorewithme.dto.out.UserDto;
+import ru.practicum.explorewithme.dto.input.create.NewUserRequestDto;
+import ru.practicum.explorewithme.dto.output.UserDto;
 import ru.practicum.explorewithme.exeption.NotFoundException;
 import ru.practicum.explorewithme.model.QUser;
 import ru.practicum.explorewithme.model.User;
